Implement search query based sharing on the UI #1015

+++ test/folder-reset.ts
@@ -5,6 +5,9 @@ declare const process: { argv: string[] };
 
 const dir = process.argv[2];
 
+if(dir === 'test'){
+  throw new Error('test folder is not allowed to be deleted');
+}
 if (fs.existsSync(dir)) {
   console.log('deleting folder:' + dir);
   fs.rmSync(dir, {recursive: true});
